fpu_control.h is glibc specific, add missing header
Currently, CVC5 does not compile with musl-libc, so it does not compile on, e.g., Alpine Linux.
The reason is that musl-libc does not provide the glibc specific header fpu_control.h.
To check whether glibc is available, features.h is required.

Including time.h in safe_print.h is also required for compilation with musl-libc.

Signed-off-by: Florian Frohn <florian.frohn@cs.rwth-aachen.de>